ramips: add support for HiWiFi 3 Pro
It's also known as board R33 as recorded in the bdinfo mtd partition.
Devices can also be known as C312B, B52 as printed on the label

This was based on work by zfdx [1]

A kernel patch for "mac-address-ascii" is backported from master branch
and reworked to support non-platform_device (wlan5g via pcie)

mediatek,port4-gmac is set so that port5 can connect to phy with id not
limited to ephy-base+5.  mediatek,ephy-base is set to 12 so that we can
use phy id 0 for the external RTL8367RB

The device has ePA for 2.4GHz signal.  However it's been observed that
for at least one device of this model enabling pa gpio pin has no
visible effect on the signal quality.  Txpower for the 2.4GHz signal can
be set to 5dBm while still exhibits good coverage

Improvement needed:

- RTL8367RB port 0 will not work if initvals were written.  Either
  switch to dsa or use linux upstream dsa jam values instead for this
  specific chip (chip num 0x6367, chip ver 0x0020)
- Rewrite NAND driver based on linux upstream
  drivers/mtd/nand/raw/mtk_nand.c.  mt7621_nand.c can be good ref

OEM firmware has a mtd partition scheme with "firmware" and
"firmware_backup".  We might also need to look into how u-boot manages
the two partitions to make sure that it won't tamper with and corrupt
the now concatenated ubi partition

Link: https://github.com/zfdx123/22.03/commit/1b9540fa04b0fbcf7a61cd9e1ccb87d47573f076
Signed-off-by: Yousong Zhou <yszhou4tech@gmail.com>